anatoliy-backend/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.1
pydantic==2.9.2
python-dateutil==2.9.0.post0
orjson==3.10.7
grpcio==1.66.1
grpcio-tools==1.66.1
